+++ app/components/Calendar/index.tsx
@@ -3,37 +3,19 @@
 import { DateBox } from '@/app/components';
 import dayjs from 'dayjs';
 import dayOfYear from 'dayjs/plugin/dayOfYear';
-import Link from 'next/link';
-import { useRouter } from 'next/navigation';
-import { Dispatch, SetStateAction, useState } from 'react';
-import { HiRefresh } from 'react-icons/hi';
+import { Dispatch, SetStateAction } from 'react';
 import { CalDate } from '../DateBox';
 import * as Styled from './Calendar.styled';
 
 type CalendarProps = {
   selectedDateId: string;
   setSelectedDateId: Dispatch<SetStateAction<string>>;
-  // TODO: if I used clerk's organization i can remove this props and get the role from the useUser
-  userRole?: string;
 };
 
 const Calendar: React.FC<CalendarProps> = ({
   selectedDateId,
   setSelectedDateId,
-  userRole,
 }) => {
-  const router = useRouter();
-  // TODO: this should be updated, using pending status from next/navigation or the fetched query.
-  const [rotate, setRotate] = useState(false);
-
-  const handleRefresh = () => {
-    setRotate(true);
-    router.refresh();
-    setTimeout(() => {
-      setRotate(false);
-    }, 3000);
-  };
-
   dayjs.extend(dayOfYear);
   const startOfWeek = dayjs().startOf('week');
   const endOfWeek = dayjs().endOf('week');
@@ -52,27 +34,16 @@ const Calendar: React.FC<CalendarProps> = ({
   }
 
   return (
-    <Styled.CalendarWrapper>
-      <div className="month-year">
-        {dayjs().format('MMMM YYYY').toUpperCase()}
-        <div className="flex gap-4">
-          {userRole === 'ADMIN' && <Link href="/admin">ADMIN</Link>}
-          <span className="icon-wrapper" onClick={() => handleRefresh()}>
-            <HiRefresh className={rotate ? 'rotate' : ''} />
-          </span>
-        </div>
-      </div>
-      <Styled.DatesWrapper>
-        {calDays.map((calDate, key) => (
-          <DateBox
-            key={key}
-            selected={selectedDateId === calDate.dateId ? true : false}
-            calDate={calDate}
-            setSelectedDateId={setSelectedDateId}
-          />
-        ))}
-      </Styled.DatesWrapper>
-    </Styled.CalendarWrapper>
+    <Styled.DatesWrapper>
+      {calDays.map((calDate, key) => (
+        <DateBox
+          key={key}
+          selected={selectedDateId === calDate.dateId ? true : false}
+          calDate={calDate}
+          setSelectedDateId={setSelectedDateId}
+        />
+      ))}
+    </Styled.DatesWrapper>
   );
 };
